+++ src/components/auth/kk-CreateAccount.css
@@ -1,0 +1,10 @@
+.fs-form{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+.create-form{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
